Stock Price Prediction
Sure, here's another Python project:

Project Name: Stock Price Prediction using Long Short-Term Memory (LSTM)

Project Description:

In this project, we will be using a Long Short-Term Memory (LSTM) model to predict the future stock prices of a particular company. We will be using the Yahoo Finance API to obtain the historical stock price data of the company and then use it to train our LSTM model. We will be using Keras, which is a high-level neural networks API, written in Python and capable of running on top of TensorFlow, CNTK, or Theano, to build and train our LSTM model.

Requirements:

    Python 3.x
    TensorFlow
    Keras
    Pandas
    NumPy
    Matplotlib

Project Steps:

    Import the required libraries and obtain the historical stock price data of the company using the Yahoo Finance API.
    Preprocess the data by selecting the columns of interest, normalizing the data, and splitting the data into training and testing sets.
    Build the LSTM model using the Keras Sequential API. Our model will have one LSTM layer, one dense layer, and one output layer.
    Compile the model and specify the loss function, optimizer, and metrics to use during training.
    Train the model on the training data and validate it on the validation data. Monitor the training progress using the training and validation loss.
    Use the trained model to make predictions on the test data and plot the actual and predicted stock prices using Matplotlib.