Automatic Email Sender
Project Description:

In this project, we will be building an Automatic Email Sender that can send emails to multiple recipients at once. We will be using the smtplib and email libraries to build and send the emails. We will also be using a CSV file to store the email addresses and message content.

Requirements:

    Python 3.x
    smtplib
    email
    pandas

Project Steps:

    Import the required libraries and read the email addresses and message content from a CSV file.
    Create a function to connect to the SMTP server and log in to the sender's email account.
    Create a function to send the emails to the recipients. In this function, we will be creating a MIME message, setting the recipients' email addresses, and adding the message content to the message.
    Loop through the recipients' email addresses and send the emails using the send_email function.
    Close the connection to the SMTP server.