ile Organizer
This program first defines the directory to be organized and the categories to sort files into. It then creates the folders for each category if they don't already exist. Finally, it loops through all the files in the directory, checks their file extension, and moves them to their respective category folder using the shutil.move function.